all done here

+++ notes/views.py
@@ -1,14 +1,19 @@
-from django.shortcuts import render, reverse, get_object_or_404, get_list_or_404
+from django.shortcuts import render, reverse, get_object_or_404, get_list_or_404, redirect
 from django.http import HttpResponse, HttpResponseRedirect
 # from django.template import loader
 from .models import note
 from uuid import uuid4 as uuid
-
+from django.contrib.auth.models import User                     # create user
+from django.contrib.auth import authenticate, login, logout     # authenticate user
+from django.contrib.auth.decorators import login_required       # ensure user's authenticated
 # Create your views here.
 
 
+
+@login_required(login_url='note:landing')
 def index(request):
-        notes = note.objects.all()
+        print(request.user.id)
+        notes = note.objects.filter(user=request.user)
         context = {'notes': notes}
         template_name = 'notes/view.html'
         return render(request, template_name, context=context)
@@ -16,7 +21,7 @@ def index(request):
         # return HttpResponse(template.render(context, request))
 
 
-
+@login_required(login_url='note:landing')
 def search(request):
         template_name = 'notes/search.html'
 
@@ -25,7 +30,7 @@ def search(request):
 
         elif request.method == 'POST':
                 query = request.POST['search']
-                notes = note.objects.filter(title=query)
+                notes = note.objects.filter(title=query, user=request.user)
                 print(notes)
                 if notes:
                         context = {
@@ -39,24 +44,20 @@ def search(request):
                         return render(request, template_name, context=context)
 
 
-
-def logout(request):
-        pass
-
-
-
-
+@login_required(login_url='note:landing')
 def bookmarked(request):
-        notes = note.objects.filter(is_bookmarked=True)
+        notes = note.objects.filter(is_bookmarked=True, user=request.user)
         context = {'notes': notes}
         template_name = 'notes/bookmarked.html'
         return render(request, template_name, context=context)
 
 
 
-
+@login_required(login_url='note:landing')
 def bookmark(request, id):
         thatNote = get_object_or_404(note, pk=id)
+        if thatNote.user != request.user:
+                return redirect('note:index')
         # bookmark the note
         print(thatNote.title)
         thatNote.is_bookmarked = not thatNote.is_bookmarked
@@ -65,7 +66,7 @@ def bookmark(request, id):
 
 
 
-
+@login_required(login_url='note:landing')
 def create(request):
         template_name = 'notes/create.html'
 
@@ -74,6 +75,7 @@ def create(request):
 
         if request.method == 'POST':
                 newNote = note()
+                newNote.user = request.user
                 newNote.title = request.POST['title']
                 newNote.content = request.POST['content']
                 newNote.id = key()
@@ -85,13 +87,84 @@ def create(request):
                 return HttpResponseRedirect(reverse('note:index'))
 
 
+@login_required(login_url='note:landing')
+def edit(request, id):
+        template_name = 'notes/edit.html'
+        thatNote = get_object_or_404(note, pk=id)
+        if thatNote.user != request.user:
+                return redirect('note:index')
+
+        elif request.method == 'GET':
+                # print(thatNote.title, thatNote.content)
+                return render(request, template_name, {'note': thatNote})
+
+        elif request.method == 'POST':
+                thatNote.title = request.POST['title']
+                thatNote.content = request.POST['content']
+                thatNote.save()
+                return HttpResponseRedirect(reverse('note:index'))
 
 
-
+@login_required(login_url='note:landing')
 def delete(request, id):
         thatNote = get_object_or_404(note, pk=id)
+        if thatNote.user != request.user:
+                return redirect('note:index')
         thatNote.delete()
         return HttpResponseRedirect(reverse('note:index'))
+
+
+
+def landing(request):
+        template_name = 'notes/landing.html'
+        return render(request, template_name)
+
+
+
+
+def register(request):
+        template_name = 'notes/auth.html'
+
+        if request.method == 'GET':
+                return render(request, template_name, {'todo': 'Register'})
+
+        elif request.method == 'POST':
+                email = request.POST['email']
+                username = request.POST['username']
+                password = request.POST['password']
+                user = User.objects.create_user(username=username, email=email, password=password)
+                # user created!
+                login(request, user)
+                # logging in
+                return redirect('note:index')
+
+
+
+def loginUser(request):
+        template_name = 'notes/auth.html'
+
+        if request.method == 'GET':
+                return render(request, template_name, {'todo': 'Login'})
+
+        elif request.method == 'POST':
+                username = request.POST['username']
+                password = request.POST['password']
+                user = authenticate(username=username, password=password)
+
+                if user is not None:
+                        login(request, user)
+                        # logging in
+                        return redirect('note:index')
+
+                # if error occurs
+                return render(request, template_name, {'todo': 'Login', 'error_message': 'invalid Email or Password'})
+
+
+
+def logoutUser(request):
+        if request.method == 'GET':
+                logout(request)
+                return redirect('note:login')
